complete review routes and logic update

+++ backend/controllers/cartController.js
@@ -4,7 +4,7 @@ const MenuItem = require("../models/menuItemModel");
 exports.getCart = async (req, res) => {
   try {
     const userId = req.user.userId;
-    const cart = await Cart.findOne({ userId })
+    const cart = await Cart.find({ userId })
       .populate("restaurantId", "name location")
       .populate("userId", "name");
     if (!cart) {
@@ -19,19 +19,17 @@ exports.getCart = async (req, res) => {
 exports.addItemToCart = async (req, res) => {
   try {
     const userId = req.user.userId;
-    const { foodId, restaurantId, quantity } = req.body;
+    const { foodId, restaurantId, quantity, cartStatus } = req.body;
     const menuItem = await MenuItem.findById(foodId).populate("restaurant");
     if (!menuItem) {
       return res.status(404).json({ message: "Menu item not found" });
     }
     const itemPrice = menuItem.price * quantity;
-    let cart = await Cart.findOne({ userId });
+    let cart = await Cart.findOne({ userId:userId, cartStatus:"ordering" });
     if (cart && cart.restaurantId.toString() !== restaurantId) {
-      return res
-        .status(409)
-        .json({
-          message: "Item from different restaurant is already added to cart",
-        });
+      return res.status(409).json({
+        message: "Item from different restaurant is already added to cart",
+      });
     }
     if (!cart) {
       cart = new Cart({
@@ -39,6 +37,7 @@ exports.addItemToCart = async (req, res) => {
         restaurantId,
         items: [],
         totalPrice: 0,
+        cartStatus,
       });
     }
     const existingItemIndex = cart.items.findIndex(
@@ -69,7 +68,7 @@ exports.addItemToCart = async (req, res) => {
       .json({ message: "Item added to cart", cart: populatedCart });
   } catch (error) {
     console.error(error);
-    res.status(500).json({ message: "Failed to add item to cart", error });
+    res.status(500).json({ message: error.message });
   }
 };
 
@@ -77,7 +76,7 @@ exports.removeItemFromCart = async (req, res) => {
   try {
     const userId = req.user.userId;
     const { foodId } = req.body;
-    const cart = await Cart.findOne({ userId })
+    const cart = await Cart.findOne({ userId: userId, cartStatus: "ordering" })
       .populate("items.foodId", "name price")
       .populate("userId", "name")
       .populate("restaurantId", "name location");
@@ -94,8 +93,10 @@ exports.removeItemFromCart = async (req, res) => {
         0
       );
       if (cart.items.length === 0) {
-        await Cart.findOneAndDelete({ userId });
-        return res.status(200).json({ message: "Cart is now empty and deleted" });
+        await Cart.findOneAndDelete({ userId: userId, cartStatus: "ordering" });
+        return res
+          .status(200)
+          .json({ message: "Cart is now empty and deleted" });
       }
       await cart.save();
       res.status(200).json({ message: "Item removed from cart", cart });
@@ -117,16 +118,17 @@ exports.updateItemQuantity = async (req, res) => {
         .status(400)
         .json({ message: "Invalid action. Use 'increment' or 'decrement'." });
     }
-    const cart = await Cart.findOne({ userId })
+    const cart = await Cart.findOne({ userId: userId, cartStatus: "ordering" })
       .populate("items.foodId", "name price")
       .populate("userId", "name")
       .populate("restaurantId", "name location");
+
     if (!cart) {
       return res.status(404).json({ message: "Cart not found" });
     }
-    const itemIndex = cart.items.findIndex(
-      (item) => item.foodId._id.toString() === foodId
-    );
+    const itemIndex = cart.items.findIndex((item) => {
+      return item.foodId._id.toString() === foodId;
+    });
     if (itemIndex === -1) {
       return res.status(404).json({ message: "Item not found in cart" });
     }
@@ -155,7 +157,10 @@ exports.updateItemQuantity = async (req, res) => {
 exports.deleteCart = async (req, res) => {
   try {
     const userId = req.user.userId;
-    const cart = await Cart.findOneAndDelete({ userId });
+    const cart = await Cart.findOneAndDelete({
+      userId: userId,
+      cartStatus: "ordering",
+    });
     if (!cart) {
       return res.status(404).send({ message: "Cart not found" });
     }
